exception.message should return a string.
## Problem

Sentry is a popular exception monitoring tool for ruby, but it fails to
capture the error raised by checkr ruby gem.

After some investigation, I found the checkr sometime set an array to
the exception message.

Here is an example

```
exception.message
=> ["First name must only contain letters, numbers, spaces, hyphens, apostrophes, periods, and commas."]
```

Usually, ruby exception's `message` method should return a string as an
abstraction of error and developer save the detail information to
internal instance variable.

## Solution

This PR is to convert the message variable to a string to avoid crashing
the monitoring tool.